+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Changed
+* Accordion: Optionele Badge in handle van Accordion Section ([#3936](https://github.com/dso-toolkit/dso-toolkit/issues/3936))
+
 ## 🕵 Release 100.3.0 - 2026-08-31
 
 ### Added